+++ packages/vue/src/components/MetadataStacked/MetadataStacked.vue
@@ -36,15 +36,15 @@ const props = defineProps({
 </script>
 
 <template>
-  <div class="MetadataStacked inline-flex flex-col flex-start gap-2.5 text-body-md">
+  <div class="MetadataStacked inline-flex flex-col flex-start gap-2 md:gap-2.5 text-body-md">
     <div
       v-if="props.location"
-      class="flex flex-row gap-2.5 items-center self-stretch"
+      class="flex flex-row gap-2 md:gap-2.5 items-start self-stretch"
     >
       <img
         v-if="props.locationIcon"
         :src="props.locationIcon"
-        class="w-4 h-4"
+        class="mt-[0.1rem] lg:mt-[0.2rem] w-6 h-6"
         alt=""
       />
       <IconLocation
@@ -55,17 +55,25 @@ const props = defineProps({
     </div>
     <div
       v-if="props.wait"
-      class="flex flex-row gap-2.5 items-center self-stretch"
+      class="flex flex-row gap-2 md:gap-2.5 items-center self-stretch"
     >
-      <IconHourglass class="MetadataStackedIcon text-jpl-red" />
+      <div class="w-6">
+        <IconHourglass class="MetadataStackedIcon text-jpl-red block mx-auto" />
+      </div>
       <span>{{ props.wait }}</span>
     </div>
     <div
       v-if="props.showMyList"
-      class="flex flex-row gap-2.5 items-center self-stretch"
+      class="flex flex-row gap-2 md:gap-2.5 items-center self-stretch"
     >
-      <IconBookmark class="MetadataStackedIcon text-gray-mid" />
+      <div class="w-6">
+        <IconBookmark class="MetadataStackedIcon text-gray-mid" />
+      </div>
       <span class="italic">Added to my list</span>
     </div>
   </div>
 </template>
+
+<style lang="scss">
+@import '@explorer-1/common/src/scss/components/MetadataStacked';
+</style>