+++ packages/vue/src/components/SearchResultCard/SearchResultCard.vue
@@ -70,6 +70,8 @@
       :wait="wait"
       :show-my-list="showMyList"
       :location-icon="locationIcon"
+      :theme="theme"
+      :description="searchDescription"
     />
     <BaseLink
       v-else
@@ -405,6 +407,17 @@ export default defineComponent({
     locationIcon: {
       type: String,
       default: ''
+    },
+    theme: {
+      type: String,
+      default: undefined
+    },
+    /**
+     * The search meta description for use as card text. Plain text only.
+     */
+    searchDescription: {
+      type: String,
+      default: undefined
     }
   },
   computed: {