+++ packages/vue/src/interfaces.ts
@@ -147,6 +147,7 @@ export interface ElasticSearchPage {
   location?: string
   locationIcon?: string
   wait?: string
+  theme?: string
   title: string
   summary?: string
   eventType?: string
@@ -155,6 +156,7 @@ export interface ElasticSearchPage {
   gradeLevels?: string
   time?: string
   targetAudience?: string
+  searchDescription?: string
 }
 
 export interface FormOption {
@@ -255,6 +257,10 @@ export interface FooterResponse {
   footer: any
 }
 
+export interface BannerAdminResponse {
+  bannerAdmin: any
+}
+
 export interface EDUSubjectArea {
   id: number
   primarySubject: EduResourcesSubject
@@ -352,6 +358,7 @@ export interface PageObject {
   waitTimeLabel?: string
   relatedPagesForMoreSection?: any
   id?: string
+  themeTypeLabel?: string
 }
 
 export interface EduResourcesSubject {