+++ packages/vue/src/templates/explore-jpl/PageSiteExploreApp/PageSiteExploreApp.stories.js
@@ -1,5 +1,5 @@
 import { BlockLinkCardCarouselData } from './../../../components/BlockLinkCarousel/BlockLinkCarousel.stories.js'
-
+import ExploreJplIconImage from '@explorer-1/common-storybook/src/story-data/images/2026-Explore-JPL-Site-Number-1-Missions-in-Flight.png'
 import PageSiteExploreApp from './PageSiteExploreApp.vue'
 
 export default {
@@ -22,6 +22,7 @@ export const ExploreAppSitePageData = {
   title: 'Sites',
   displayLabel: 'Explore JPL',
   showShareLinks: false,
+  themeTypeLabel: 'Missions That Changed the World',
   thumbnailImage: {
     original: 'https://picsum.photos/512/288',
     src: {
@@ -39,7 +40,9 @@ export const ExploreAppSitePageData = {
   location: 'Cafe 186',
   waitTime: '60',
   waitTimeLabel: '60 minute wait',
-  locationIcon: '',
+  mapIcon: {
+    url: ExploreJplIconImage
+  },
   body: [
     {
       blockType: 'RichTextBlock',